add option of flag checking
