+++ CalPatRec/inc/DeltaFinder_types.hh
@@ -9,6 +9,9 @@ namespace fhicl {
   class ParameterSet;
 };
 
+#include "TObject.h"
+#include "TClonesArray.h"
+
 #include "fhiclcpp/types/Atom.h"
 #include "fhiclcpp/types/Table.h"
 
@@ -20,6 +23,10 @@ namespace fhicl {
 #include "Offline/TrackerGeom/inc/Tracker.hh"
 
 #include "Offline/Mu2eUtilities/inc/McUtilsToolBase.hh"
+
+#include "Offline/CalPatRec/inc/DeltaFinder_structures.hh"
+#include "Offline/CalPatRec/inc/DeltaSeed.hh"
+#include "Offline/CalPatRec/inc/DeltaCandidate.hh"
 
 namespace mu2e {
   class Panel;
@@ -51,254 +58,6 @@ namespace mu2e {
       fhicl::Table<McUtilsToolBase::Config> mcUtils{fhicl::Name("mcUtils"       ), fhicl::Comment("MC Diag plugin") };
     };
 //-----------------------------------------------------------------------------
-// intersection of the two hit wires
-//-----------------------------------------------------------------------------
-    struct Intersection_t {
-      double     x;                        // x-coordinate of the intersection point
-      double     y;                        // y-coordinate of the intersection point
-      double     z;                        // y-coordinate of the intersection point
-      double     wd1;                      // distance btw the 1st hit and the intersection
-      double     wd2;                      // distance btw the 2nd hit and the intersection
-    };
-
-    struct DeltaCandidate;
-    struct DeltaSeed;
-
-    enum {
-      kNStations      = StrawId::_nplanes/2,   // number of tracking stations
-      kNFaces         = StrawId::_nfaces*2 ,   // N(faces) per station (4)
-      kNPanelsPerFace = StrawId::_npanels/2    // = 3
-    };
-
-    struct HitData_t {
-      const ComboHit*         fHit;
-      DeltaSeed*              fSeed;           // nullptr if not associated...
-      int                     fNSecondHits;
-      int                     fDeltaIndex;
-      float                   fChi2Min;
-      float                   fSigW2;    // cached resolution^2 along the wire
-      float                   fCorrTime; // cache hit corrected time
-
-      HitData_t(const ComboHit* Hit) {
-        fHit         = Hit;
-        fChi2Min     = 99999.0;
-        float sigw   =  Hit->posRes(ComboHit::wire);
-        fSigW2       = sigw*sigw;  // to be used to calculate chi2...
-        fSeed        = nullptr;
-        fNSecondHits =  0;
-        fDeltaIndex  = -1;
-        fCorrTime    = Hit->correctedTime();
-      }
-
-      int Used() const { return (fSeed != nullptr) ; }
-    };
-
-    struct PanelZ_t {
-      int                              fNHits  ; // guess, total number of ComboHits
-      std::vector<HitData_t>           fHitData;
-      const Panel*                     fPanel;      // backward pointer to the tracker panel
-      double                           wx;          // direction cosines of the wires, assumed to be all the same
-      double                           wy;
-      double                           phi;         // phi angle of the wire
-      double                           z;           //
-      float                            tmin;        // for hits stored on this panel
-      float                            tmax;
-    };
-
-//-----------------------------------------------------------------------------
-// diagnostics structure
-//-----------------------------------------------------------------------------
-    struct McPart_t {
-      const SimParticle*            fSim;        // type undefined here
-      const DeltaCandidate*         fDelta;
-      std::vector<const HitData_t*> fListOfHits;
-      int                           fFirstStation;
-      int                           fLastStation;
-      int                           fID;         // SimParticle::id().asInt()
-      int                           fPdgID;
-      int                           fNHitsCE;
-      int                           fNHitsDelta; // number of hits associated with reconstructed delta electrons
-      float                         fTime;
-      float                         fHitTMin;    // min and max times of the straw hits
-      float                         fHitTMax;
-      float                         fStartMom;
-
-      McPart_t(const SimParticle* Sim = NULL) {
-        fSim          = Sim;
-        fDelta        = NULL;
-        fID           = -1;
-        fPdgID        = 0;
-        fNHitsDelta   = 0;
-        fNHitsCE      = 0;
-        fFirstStation = 999;
-        fLastStation  = -1;
-        fStartMom     = -1;
-        fTime         = 1.e6;                 // at initialization, make it absurd
-        fHitTMin      = 1.e6;
-        fHitTMax      = -1.e6;
-      }
-
-      ~McPart_t() { fListOfHits.clear(); }
-
-      int NHits() { return fListOfHits.size(); }
-
-      float Momentum () { return fStartMom; }
-      float Time     () { return fTime; }
-      float HitDt    () { return fHitTMax-fHitTMin; }
-    };
-
-    class DeltaSeed {
-    public:
-      int          fIndex;            // index within the station
-      int          fStation;          // station with seed stereo hit
-      int          fType;             // defines indices of the two faces used for preseed seach
-      int          fGood;             // <killer number> if not to be used - what about 0 ?
-      int          fNFacesWithHits;
-      int          fNHits;            // total number of combo hits
-      int          fNStrawHits;       // total number of straw hits
-      int          fNHitsCE;          // number of associated CE hits
-
-      int          fSFace[2];         // faces making the stereo seed
-      float        fChi21;            // chi2's of the two initial hits, also stored in hit data
-      float        fChi22;
-                                      // 0: used in recovery
-      int          fFaceProcessed[kNFaces];
-      HitData_t*   hitData       [kNFaces];
-      McPart_t*    fMcPart       [kNFaces];
-                                      // XY coordinate sums
-      double       fSx;
-      double       fSy;
-      double       fSnx2;
-      double       fSnxny;
-      double       fSny2;
-      double       fSnxnr;
-      double       fSnynr;
-      float        fSumEDep;             // sum over the straw hits
-
-      XYZVectorF   CofM;                 // COG
-      float        fPhi;                 // cache to speed up the phi checks
-      float        fZ;                   // Z-coordinate of the center of the corresponding station
-
-      double       fSumT;
-
-      float        fMinHitTime;          // min and max times of the included hits
-      float        fMaxHitTime;
-      McPart_t*    fPreSeedMcPart[2];    // McPart_t's for initial intersection
-      int          fDeltaIndex;
-                                         // chi2's
-      float        fChi2All;
-      float        fChi2Perp;
-
-      DeltaSeed ();
-      DeltaSeed (int Index, int Station, int Face0, HitData_t* Hd0, int Face1, HitData_t* Hd1);
-      ~DeltaSeed() {}
-
-      int              Station ()         { return fStation; }
-      int              Index   ()         { return fIndex; }
-      int              SFace(int I)       { return fSFace[I]; }
-      float            Chi2TotN()         { return (fChi21+fChi22)/fNHits; }
-      float            Chi2Tot ()         { return (fChi21+fChi22); }
-      float            Chi2All ()         { return fChi2All; }
-      float            Chi2Perp()         { return fChi2Perp; }
-      float            Chi2PerpN()        { return fChi2Perp/fNHits; }
-      float            Chi2AllN ()        { return fChi2All/fNHits ; }
-      HitData_t*       HitData (int Face) { return hitData[Face]; } // no boundary check !
-      int              NHits   ()         { return fNHits; }
-      int              NHitsCE ()         { return fNHitsCE; }
-      int              NStrawHits()       { return fNStrawHits; }
-      float            SumEDep ()         { return fSumEDep ; }
-      float            EDep    ()         { return fSumEDep/fNStrawHits ; }
-      int              MCTruth ()         { return (fPreSeedMcPart[0] != NULL) && (fPreSeedMcPart[0] == fPreSeedMcPart[1]) ; }
-      bool             Used    ()         { return (fDeltaIndex >= 0); }
-      int              Good    ()         { return (fGood       >= 0); }
-      float            TMean   ()         { return fSumT/fNStrawHits;  }
-      float            Z       ()         { return fZ;  }
-
-      float            Phi     () const   { return fPhi; }
-//-----------------------------------------------------------------------------
-// drift time can't be < 0
-// fMaxTime < particle T0 < fMinTime
-//-----------------------------------------------------------------------------
-      float            Time      () { return (fMinHitTime+fMaxHitTime)/2;    }
-      float            MinHitTime() { return fMinHitTime; }
-      float            MaxHitTime() { return fMaxHitTime; }
-//-----------------------------------------------------------------------------
-// assumed range of allowed hit times for this seed
-//-----------------------------------------------------------------------------
-      float            T0Min     () { return (fMinHitTime+fMaxHitTime)/2-20; }
-      float            T0Max     () { return (fMinHitTime+fMaxHitTime)/2+20; }
-//-----------------------------------------------------------------------------
-// less trivial functions
-//-----------------------------------------------------------------------------
-      void             AddHit         (HitData_t* Hd, int Face);
-      void             ReplaceFirstHit(HitData_t* Hd);
-      void             CalculateCogAndChi2(double SigmaR2);
-    };
-
-    struct DeltaCandidate {
-    public:
-      int                   fIndex;                 // >= 0: index, <0: -1000-index merged
-      int                   fFirstStation;
-      int                   fLastStation;
-      DeltaSeed*            seed   [kNStations];
-      float                 dxy    [kNStations];   // used only for diagnostics
-      float                 fT0Min [kNStations];   // acceptable hit times (no need to account for the drift time!)
-      float                 fT0Max [kNStations];
-      XYZVectorF            CofM;
-      float                 phi;
-      int                   n_seeds;
-      McPart_t*             fMcPart;
-      int                   fNHits;                // n(combo hits)
-      int                   fNStrawHits;
-      int                   fNHitsMcP;             // N combo hits by the "best" particle"
-      int                   fNHitsCE;
-      float                 fSumEDep;              //
-                                                   // LSQ sums
-      double                fSx;
-      double                fSy;
-      double                fSnx2;
-      double                fSnxny;
-      double                fSny2;
-      double                fSnxnr;
-      double                fSnynr;
-                                                   // LSQ sumz for TZ
-      double                fSt;
-      double                fSz;
-      double                fSt2;
-      double                fStz;
-      double                fSz2;
-
-      double                fT0;
-      double                fDtDz;
-      double                fSigT0;
-//-----------------------------------------------------------------------------
-//
-//-----------------------------------------------------------------------------
-      DeltaCandidate();
-      DeltaCandidate(int Index, DeltaSeed* Seed, int Station);
-
-      int        Active               () const { return (fIndex >= 0); }
-      int        Index                () const { return fIndex ; }
-      int        NSeeds               () const { return n_seeds; }
-      int        NHits                () const { return fNHits; }
-      int        NHitsMcP             () const { return fNHitsMcP; }
-      int        NStrawHits           () const { return fNStrawHits; }
-      DeltaSeed* Seed            (int I) const { return seed[I]; }
-      bool       StationUsed     (int I) const { return (seed[I] != NULL); }
-      float      T0Min           (int I) const { return fT0Min[I]; }
-      float      T0Max           (int I) const { return fT0Max[I]; }
-      float      Time            (int I) const { return (fT0Max[I]+fT0Min[I])/2.; }
-      int        LastStation          () const { return fLastStation ; }
-      int        FirstStation         () const { return fFirstStation; }
-      float      EDep                 () const { return fSumEDep/fNStrawHits; }
-      float      FBest                () const { return float(fNHitsMcP)/fNHits; }
-
-      void       AddSeed            (DeltaSeed*      Ds   , int Station);
-      void       MergeDeltaCandidate(DeltaCandidate* Delta, int PrintErrors);
-
-      void       SetIndex(int Index) { fIndex = Index; }
-    };
-//-----------------------------------------------------------------------------
 // data structure passed to the diagnostics plugin
 //-----------------------------------------------------------------------------
     struct Data_t {
@@ -308,21 +67,43 @@ namespace mu2e {
       art::InputTag                 chfCollTag;
 
       art::InputTag                 sdmcCollTag;
-      std::vector<DeltaSeed*>       listOfSeeds[kNStations]; // seeds with the first station being this
+
+      TClonesArray*                 listOfSeeds[kNStations]; // seeds with the first station being this
+
       std::vector<DeltaCandidate>   listOfDeltaCandidates;
 
       PanelZ_t                      oTracker   [kNStations][kNFaces][kNPanelsPerFace];
       int                           stationUsed[kNStations];
 
-      int                           nseeds;
+      int                           fNSeeds;
       int                           nseeds_per_station[kNStations];
       const ComboHitCollection*     chcol;
-      StrawHitFlagCollection*       chfcol;       // output combohit flags
+      const StrawHitFlagCollection* chfColl;            // input  combohit flags
+      StrawHitFlagCollection*       outputChfColl;      // output combohit flags
       const TimeClusterCollection*  tpeakcol;
       int                           debugLevel;   // printout level
 
       DeltaCandidate* deltaCandidate(int I)              { return &listOfDeltaCandidates[I]; }
-      DeltaSeed*      deltaSeed     (int Station, int I) { return listOfSeeds[Station][I]; }
+      DeltaSeed*      deltaSeed     (int Station, int I) {
+        return (DeltaSeed*) listOfSeeds[Station]->UncheckedAt(I);
+      }
+
+
+      void InitEvent(const art::Event* Evt, int DebugLevel);
+
+      int NSeedsTot() { return fNSeeds; }
+
+      int NSeeds(int Station) {
+        return listOfSeeds[Station]->GetEntriesFast();
+      }
+
+      DeltaSeed*      NewDeltaSeed(int Station, int Face0, HitData_t* Hd0, int Face1, HitData_t* Hd1) {
+        int loc = nseeds_per_station[Station];
+        DeltaSeed* ds = new ((*listOfSeeds[Station])[loc]) DeltaSeed(loc,Station,Face0,Hd0,Face1,Hd1);
+        nseeds_per_station[Station] += 1;
+        fNSeeds++;
+        return ds;
+      }
 
       void printHitData       (HitData_t*      HitData, const char* Option = "");
       void printDeltaSeed     (DeltaSeed*      Seed   , const char* Option = "");
@@ -330,7 +111,7 @@ namespace mu2e {
     };
 
 //-----------------------------------------------------------------------------
-// finally, utility functions
+// finally, utility functions still used by the diag tool
 //-----------------------------------------------------------------------------
     int findIntersection(const HitData_t* Hd1, const HitData_t* Hd2, Intersection_t* Result);
   }
