+++ CalPatRec/src/classes.h
@@ -1,0 +1,10 @@
+//
+// headers needed when genreflex creates root dictionaries
+// for objects written to art files
+//
+
+#define ENABLE_MU2E_GENREFLEX_HACKS
+
+#include "Offline/CalPatRec/inc/DeltaSeed.hh"
+
+#undef ENABLE_MU2E_GENREFLEX_HACKS
