+++ CalPatRec/src/classes_def.xml
@@ -1,0 +1,11 @@
+<!--  -->
+<!--  Only include art::Wrapper lines for objects that we would like to put into the event -->
+<!--  Include the non-wrapper lines for all objects on the art::Wrapper lines and     -->
+<!--  for all objects inside of those objects.                                        -->
+<!--  -->
+
+<lcgdict>
+
+<class name="mu2e::DeltaSeed"/>
+
+</lcgdict>
