+++ CalPatRec/test/flagBkgHits_test3.fcl
@@ -1,0 +1,73 @@
+# -*- mode:tcl -*-
+# variables which need defaults:
+# -----------------------------
+# :DEFAULT: TRIGGER_PATH=p6
+#------------------------------------------------------------------------------
+#include "Offline/fcl/minimalMessageService.fcl"
+#include "Offline/fcl/standardProducers.fcl"
+#include "Offline/fcl/standardServices.fcl"
+BEGIN_PROLOG
+  INPUT_MODULE  @protect_ignore: RootInput        # choices: EmptyEvent / RootInput
+  TRIGGER_PATHS @protect_ignore: [ p1       ]     # defined below
+END_PROLOG
+
+#include "Stntuple/fcl/stntuple.fcl"
+
+process_name : TestDeltaFinder
+
+physics : {
+
+    producers : { @table::Reconstruction.producers
+        DeltaFinder: { @table::CalPatRec.producers.DeltaFinder
+            diagLevel                     : 0 ## 1 ##
+            debugLevel                    : 0 ## 11 ## 1 ##
+            useTimePeaks                  : 0
+            diagPlugin : { @table::CalPatRec.producers.DeltaFinder.diagPlugin
+                mcDiag                    : true
+                printElectrons            : 0 ## 1 ##
+                printElectronsHits        : 0 ## 1 ##
+                printElectronsMinMom      : 0.
+                printElectronsMaxMom      : 20.
+                printElectronsMinNHits    : 1   ## 30
+                printElectronsMaxFReco    : 1.1 ## 0.5
+            }
+        }
+
+        DeltaFinderAna: { @table::CalPatRec.producers.DeltaFinderAna
+            debugLevel     : 0 ## 1
+            # tcCollTag    : "DeltaFinder"
+            chCollTag      : "makePH"
+            chfCollTag     : "FlagBkgHits:ComboHits"
+            shfCollTag     : "FlagBkgHits:StrawHits"
+            sdmcCollTag    : "compressDigiMCs"
+        }
+    }
+
+    filters   : { @table::Reconstruction.filters   }
+    analyzers : { @table::stntuple.analyzers       }
+
+    p1: [ @sequence::TrkHitReco.PrepareHits, DeltaFinder, DeltaFinderAna ]
+    p2: [ @sequence::TrkHitReco.PrepareHits, DeltaFinderAna ]
+
+    trigger_paths : [ p2            ]
+
+    e1            : [               ] ## MuHitDisplay  ] # defaultOutput ]
+    end_paths     : [ e1            ]
+}
+
+physics.producers.makePH.StrawHitSelectionBits     : [ "TimeSelection"] ##, "EnergySelection" ] ## default
+physics.producers.makePH.MaxDS                     : 4
+# physics.producers.makePH.MinimumRadius             : 370.
+## physics.producers.makePH.MaximumRadius             : 680.
+physics.producers.FlagBkgHits.TNTClustering.SignalMask : ["TimeSelection", "RadiusSelection" ] ## , "EnergySelection" ]
+#
+physics.producers.DeltaFinderAna.printElectronHits : 1
+#------------------------------------------------------------------------------
+# uncomment to write the DST out
+#------------------------------------------------------------------------------
+# physics.out : [detectorOutput]
+# outputs.detectorOutput.fileName           : "sim.murat.bbb.ccc.read_reco_stn_tcn.art"
+services.TFileService.fileName            : "flagBkgHits_test3.root"
+#------------------------------------------------------------------------------
+# everything below is appended by murat/scripts/submit_mu2e_job
+#------------------------------------------------------------------------------
